Color config dialog change + minor bugfixes (#64)
1) Separate settings for screen & print channel/spectrum colors
2) eliminated warning ...helper.h:111:2: warning: extra ‘;’ [-Wpedantic]
3) eliminated warning ...control.cpp:724:26: warning: overflow in implicit constant conversion [-Woverflow]
4) min width for Config dialog Icons widget
5) wrong type cast removed
6) print spectrum magnitude only if spectrum is enabled

+++ openhantek/src/hantek/control.cpp
@@ -721,7 +721,7 @@ unsigned int Control::updateSamplerate(unsigned int downsampler,
       else { // Downsampling factors 3 and 4 are not supported
         samplerateId = 3;
         downsampler = 5;
-        downsamplerValue = 0xffff;
+        downsamplerValue = (short int)0xffff;
       }
     } else {
       // For any dividers above the downsampling factor can be set directly
